app/config/config.py
from dotenv import load_dotenv
import os
load_dotenv()
DATABASE_URL = os.getenv("DATABASE_URL")
SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM= os.getenv("ALGORITHM")
EXPIRY_TIME = int(os.getenv("EXPIRY_TIME", "15"))
